src/main/webapp/app/entities/societaire/delete/societaire-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ISocietaire } from '../societaire.model';
import { SocietaireService } from '../service/societaire.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './societaire-delete-dialog.component.html',
})
export class SocietaireDeleteDialogComponent {
  societaire?: ISocietaire;

  constructor(protected societaireService: SocietaireService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.societaireService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
